+++ easy-ui-react/src/DataGrid/constants.ts
@@ -1,0 +1,3 @@
+export const EXPAND_COLUMN_KEY = "__ezui-column-key-expand__";
+export const ACTIONS_COLUMN_KEY = "__ezui-column-key-actions__";
+export const DEFAULT_MAX_ROWS = 999;